Update README.md
Refine the definition of the list a bit. I give people the benefit of the doubt, on the list, as in life. That will lead to mistakes, but I think being on the list will lead to added scrutiny, and I will remove people I hear a bunch of complaints about, which is probably worse than never being on the list.