Update games.md
Thanks greatly to Craig Grannell, I've tightened the standards for this list, and changed the cutoff for the amount of artistic violent in a game and bad behavior in a company to the Legend of Zelda: A Link to the Past by Nintendo. I think in general that stricter standards will make these lists more useful. They also soften the hyperbolic implications of non-inclusion.