Update games.md
In order to keep things consistent, and as useful as possible, I've followed the original format of providing links to the iOS version and putting games of particular merit not available on iOS in a separate section below.

If there is demand for it, I would consider doing something like the developer list where links to different version are provided after each name. This is extra work to create and maintain, so I will only do it if there's a real demand for it.